chore: directory structure has changed, user email is displayed when loggedin
