+++ src/components/ErrorPage/index.tsx
@@ -1,0 +1,3 @@
+export default function ErrorPage() {
+  return <p>Sorry, something went wrong</p>;
+}
