+++ src/pages/error/index.tsx
@@ -1,3 +1,1 @@
-export default function ErrorPage() {
-  return <p>Sorry, something went wrong</p>;
-}
+export { default } from "@/components/ErrorPage";
