+++ src/pages/score/index.tsx
@@ -1,8 +1,29 @@
-import Layout from "@/components/Layout";
-export default function ScorePage() {
-  return (
-    <Layout>
-      <div>Score Page</div>
-    </Layout>
-  );
+import type { User } from "@supabase/supabase-js";
+import type { GetServerSidePropsContext } from "next";
+import { createClient } from "@/utils/supabase/server-props";
+import ScorePage from "@/components/ScorePage";
+
+export default function Score({ user }: { user: User }) {
+  return <ScorePage user={user} />;
+}
+
+export async function getServerSideProps(context: GetServerSidePropsContext) {
+  const supabase = createClient(context);
+
+  const { data, error } = await supabase.auth.getUser();
+
+  if (error || !data) {
+    return {
+      redirect: {
+        destination: "/",
+        permanent: false,
+      },
+    };
+  }
+
+  return {
+    props: {
+      user: data.user,
+    },
+  };
 }
