+++ src/utils/supabase/static-props.ts
